feat(gm): native query cluster + goto/summon + missionAssign (11 commands)
Bucket A continued — native gm* cell methods (no .-prefix), FanMMORPG behavior
as reference. All cell-side:

Query cluster (report via the feedback channel):
- listAbilities (123), gmShowFlag (125), gmGetMobAttribute (127),
  gmShowMobCount (128), gmDebugMobData (180)
- gmMissionList (113), gmMissionListFull (114), gmMissionDetails (115)

Mutations (reuse existing primitives):
- gmGoto (160) — teleport caller to a target entity (numeric id, same-space)
- gmSummon (161) — move target to caller (player: TeleportPlayer; NPC: grid)
- gmMissionAssign (109) — accept_mission via the mission-def first step/objectives

Also split the oversized gm/tests.rs (1060 lines) into a gm/tests/ directory by
family (give/stats/missions/travel/world/query + shared helpers in mod.rs); all
files now <= ~210 lines.

dispatch-table tally 24->35 DONE / 49->38 ADAPT; ADAPT plan updated. 51 gm tests
pass; fmt + clippy clean.

Remaining Bucket A (need new base-side infra, next commit): gmSpawnByCmd (185,
cell<->base template round-trip), gmGiveExpertise (139) + gmGiveAppliedSciencePoints
(140, new CellToBaseMsg + crafting persistence).
